backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Audio Processing
deepfilternet==0.5.6
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Queue System
celery==5.3.4
redis==5.0.1

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Security
pyclamd==0.4.0  # ClamAV integration (optional)

# Utilities
python-dotenv==1.0.0
pillow==10.1.0
psutil==5.9.6

# CORS
fastapi-cors==0.0.6

# Monitoring and Logging
python-json-logger==2.0.7

# Background tasks
apscheduler==3.10.4
